faiss-pic-demo/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
faiss-cpu>=1.7.0
pillow>=8.3.0
streamlit>=1.10.0
pathlib